Renamed benefit files.
