+++ gettsim/benefits/elterngeld.py
@@ -4,7 +4,7 @@ import numpy as np
 from dateutil import relativedelta
 
 from gettsim.apply_tax_funcs import apply_tax_transfer_func
-from gettsim.benefits.arbeitslosengeld import proxy_net_wage_last_year
+from gettsim.benefits.arbeitsl_geld import proxy_net_wage_last_year
 
 
 def elterngeld(
