+++ gettsim/tax_transfer.py
@@ -1,10 +1,10 @@
 from gettsim.apply_tax_funcs import apply_tax_transfer_func
-from gettsim.benefits.alg2 import alg2
-from gettsim.benefits.arbeitslosengeld import ui
+from gettsim.benefits.arbeitsl_geld import ui
+from gettsim.benefits.arbeitsl_geld_2 import alg2
 from gettsim.benefits.benefit_checks import benefit_priority
 from gettsim.benefits.elterngeld import elterngeld
-from gettsim.benefits.kiz import kiz
-from gettsim.benefits.unterhaltsvorschuss import uhv
+from gettsim.benefits.kinderzuschlag import kiz
+from gettsim.benefits.unterhalt import uhv
 from gettsim.benefits.wohngeld import wg
 from gettsim.checks import check_boolean
 from gettsim.incomes import disposable_income
