+++ gettsim/tests/test_arbeitslosengeld_ui.py
@@ -3,7 +3,7 @@ import pytest
 from pandas.testing import assert_series_equal
 
 from gettsim.apply_tax_funcs import apply_tax_transfer_func
-from gettsim.benefits.arbeitslosengeld import ui
+from gettsim.benefits.arbeitsl_geld import ui
 from gettsim.config import ROOT_DIR
 from gettsim.policy_for_date import get_policies_for_date
