+++ gettsim/tests/test_kiz.py
@@ -3,7 +3,7 @@ import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal
 
-from gettsim.benefits.kiz import kiz
+from gettsim.benefits.kinderzuschlag import kiz
 from gettsim.config import ROOT_DIR
 from gettsim.policy_for_date import get_policies_for_date
