+++ gettsim/tests/test_uhv.py
@@ -3,7 +3,7 @@ import pandas as pd
 import pytest
 from pandas.testing import assert_series_equal
 
-from gettsim.benefits.unterhaltsvorschuss import uhv
+from gettsim.benefits.unterhalt import uhv
 from gettsim.config import ROOT_DIR
 from gettsim.policy_for_date import get_policies_for_date
